Adds quantization matrices
This adds compensation for the magnitude of the basis functions. Since we're
now using fixed lapping, the magnitudes are always the same for a certain block
size, making it possible to compensate. We now also have per-coefficient
quantization matrices. The same quantization matrix is used for quantizing
and for weighting the coefficient distortion in the block size RDO. We
also introduce compensation for the scaling of the Haar DC basis functions.