Simplifies the Q15 adaptation
We move the floor subtract/add to the tmp value. Also, we add an offset such
that the floor can actually reach 1/ft even for the first and last symbol.
We also get rid of the rounding for the initial adaptation since it didn't
do much and only has an effect on the first and last probabilities.